+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v1/beacon/PostBlindedBlock.java
@@ -16,7 +16,6 @@ package tech.pegasys.teku.beaconrestapi.handlers.v1.beacon;
 import static tech.pegasys.teku.beaconrestapi.handlers.v1.beacon.MilestoneDependentTypesUtil.getSchemaDefinitionForAllSupportedMilestones;
 import static tech.pegasys.teku.beaconrestapi.handlers.v1.beacon.MilestoneDependentTypesUtil.slotBasedSelector;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_ACCEPTED;
-import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_INTERNAL_SERVER_ERROR;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_OK;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_SERVICE_UNAVAILABLE;
 import static tech.pegasys.teku.infrastructure.http.RestApiConstants.TAG_BEACON;
@@ -28,9 +27,7 @@ import java.util.Optional;
 import tech.pegasys.teku.api.DataProvider;
 import tech.pegasys.teku.api.SyncDataProvider;
 import tech.pegasys.teku.api.ValidatorDataProvider;
-import tech.pegasys.teku.infrastructure.restapi.endpoints.AsyncApiResponse;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.EndpointMetadata;
-import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiEndpoint;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequest;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBlockContainer;
@@ -38,11 +35,8 @@ import tech.pegasys.teku.spec.datastructures.validator.BroadcastValidationLevel;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionCache;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitions;
 
-public class PostBlindedBlock extends RestApiEndpoint {
+public class PostBlindedBlock extends AbstractPostBlock {
   public static final String ROUTE = "/eth/v1/beacon/blinded_blocks";
-
-  private final ValidatorDataProvider validatorDataProvider;
-  private final SyncDataProvider syncDataProvider;
 
   public PostBlindedBlock(
       final DataProvider dataProvider,
@@ -60,9 +54,7 @@ public class PostBlindedBlock extends RestApiEndpoint {
       final SyncDataProvider syncDataProvider,
       final Spec spec,
       final SchemaDefinitionCache schemaDefinitionCache) {
-    super(getEndpointMetaData(spec, schemaDefinitionCache));
-    this.validatorDataProvider = validatorDataProvider;
-    this.syncDataProvider = syncDataProvider;
+    super(validatorDataProvider, syncDataProvider, createMetadata(spec, schemaDefinitionCache));
   }
 
   @Override
@@ -77,22 +69,10 @@ public class PostBlindedBlock extends RestApiEndpoint {
     request.respondAsync(
         validatorDataProvider
             .submitSignedBlindedBlock(requestBody, BroadcastValidationLevel.NOT_REQUIRED)
-            .thenApply(
-                result -> {
-                  if (result.isSuccessful()) {
-                    return AsyncApiResponse.respondWithCode(SC_OK);
-                  }
-                  if (result.isNotImportedDueToInternalError()) {
-                    return AsyncApiResponse.respondWithError(
-                        SC_INTERNAL_SERVER_ERROR,
-                        "An internal error occurred, check the server logs for more details.");
-                  } else {
-                    return AsyncApiResponse.respondWithCode(SC_ACCEPTED);
-                  }
-                }));
+            .thenApply(this::processSendSignedBlockResult));
   }
 
-  private static EndpointMetadata getEndpointMetaData(
+  private static EndpointMetadata createMetadata(
       final Spec spec, final SchemaDefinitionCache schemaDefinitionCache) {
     return EndpointMetadata.post(ROUTE)
         .operationId("publishBlindedBlock")